Explicit casts
serialize.c:
    Added sanity check.

proto_kvlds_server.c:
    We already checked that nkeys <= UINT32_MAX

proto_s3_client.c:
    We already checked that buflen and maxlen are both
        <= PROTO_S3_MAXLEN      (= 0x80000000)

dynamodb_kv.c:
    Checked 2 lines above.

+++ lib/dynamodb/dynamodb_kv.c
@@ -260,7 +260,7 @@ dynamodb_kv_extractv(const uint8_t * inbuf, size_t inlen,
 	/* Record the size of the returned value. */
 	if (vlen >= (uint32_t)(-1))
 		goto novalue1;
-	*outlen = vlen;
+	*outlen = (uint32_t)vlen;
 
 	/* Success! */
 	return (0);
